# Performance improvements and simplifications for fixed size row-based rolling windows