--- a/cpp/src/rolling/detail/rolling_utils.cuh
+++ b/cpp/src/rolling/detail/rolling_utils.cuh
@@ -21,7 +21,7 @@
 
 #include <cuda/functional>
 
-namespace cudf {
+namespace CUDF_EXPORT cudf {
 
 namespace detail::rolling {
 
@@ -118,4 +118,4 @@ template <direction Direction, typename Grouping>
     cudf::size_type{0}, fixed_window_clamper<Direction, Grouping>{grouper, delta});
 }
 }  // namespace detail::rolling
-}  // namespace cudf
+}  // namespace CUDF_EXPORT cudf