## Description

Having introduced benchmarks of rolling window performance in #17613, we can now look at what happens when cleaning up some the implementation.

### Breaking change

Previously, the most general "variable-sized" rolling window function applied a transform iterator (sometimes multiple times) to the user-provided `preceding` and `following` columns. We now _require_, and document, that the user provides values that are always in bounds.

This simplifies the book-keeping, and allows us to remove some (now redundant) checks from the rolling kernel.

### Performance improvements

In various places, we accept fixed size integers as row-based offsets for rolling windows. We used to materialise these into columns representing the clamped `preceding` and `following` columns. However, we did so inconsistently with respect to which side we would clamp (after the change to allow negative window offsets).

To fix this, rationalise this clamping in one place, and just use a transform iterator. We now do this consistently for both grouped and ungrouped fixed-size windows.

### Benchmark results

Comparing da4acccf6b76a93b1e9395b51950a17c45da99c3 (the tip of #17613) and this branch:

### nvbench compare output

Headline numbers:

- group-aware fixed rolling windows between 10 and 30% faster, use 30% less memory
- ungrouped fixed rolling windows between 10 and 50% faster, use 50% less memory
- ungrouped variable rolling windows basically no change (so no regression)

<details> <summary> All the data </summary> 

latest/_deps/nvbench-src/scripts/nvbench_compare.py da4acccf6b76a93b1e9395b51950a17c45da99c3.json c3973aa36c373e013c6c90f8d7bd58fe0aed6880.json 
['da4acccf6b76a93b1e9395b51950a17c45da99c3.json', 'c3973aa36c373e013c6c90f8d7bd58fe0aed6880.json']

# row_grouped_rolling_sum

## [0] NVIDIA RTX A6000

|  T  |  num_rows  |  preceding_size  |  following_size  |  min_periods  |  cardinality  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |         Diff |   %Diff |  Status  |
|-----|------------|------------------|------------------|---------------|---------------|------------|-------------|------------|-------------|--------------|---------|----------|
| I32 |    2^14    |        1         |        2         |       1       |      10       | 123.556 us |       3.90% | 111.199 us |       4.09% |   -12.357 us | -10.00% |   FAIL   |
| I32 |    2^28    |        1         |        2         |       1       |      10       |  24.945 ms |       0.13% |  17.126 ms |       0.26% | -7819.548 us | -31.35% |   FAIL   |
| I32 |    2^14    |        10        |        2         |       1       |      10       | 122.737 us |       4.09% | 109.055 us |       4.40% |   -13.682 us | -11.15% |   FAIL   |
| I32 |    2^28    |        10        |        2         |       1       |      10       |  24.944 ms |       0.18% |  17.151 ms |       0.20% | -7793.093 us | -31.24% |   FAIL   |
| I32 |    2^14    |        1         |        2         |       1       |      100      | 119.947 us |       4.07% | 108.432 us |       5.18% |   -11.515 us |  -9.60% |   FAIL   |
| I32 |    2^28    |        1         |        2         |       1       |      100      |  24.954 ms |       0.15% |  17.149 ms |       0.20% | -7805.558 us | -31.28% |   FAIL   |
| I32 |    2^14    |        10        |        2         |       1       |      100      | 123.919 us |       3.85% | 109.812 us |       4.63% |   -14.107 us | -11.38% |   FAIL   |
| I32 |    2^28    |        10        |        2         |       1       |      100      |  24.955 ms |       0.14% |  17.173 ms |       0.18% | -7782.547 us | -31.19% |   FAIL   |
| I32 |    2^14    |        1         |        2         |       1       |    1000000    | 121.896 us |       4.59% | 109.115 us |       4.57% |   -12.781 us | -10.49% |   FAIL   |
| I32 |    2^28    |        1         |        2         |       1       |    1000000    |  25.291 ms |       0.15% |  17.460 ms |       0.21% | -7830.713 us | -30.96% |   FAIL   |
| I32 |    2^14    |        10        |        2         |       1       |    1000000    | 126.450 us |       3.65% | 111.787 us |       3.37% |   -14.662 us | -11.60% |   FAIL   |
| I32 |    2^28    |        10        |        2         |       1       |    1000000    |  25.295 ms |       0.13% |  17.593 ms |       0.19% | -7701.923 us | -30.45% |   FAIL   |
| I32 |    2^14    |        1         |        2         |       1       |   100000000   | 122.268 us |       4.42% | 109.338 us |       5.17% |   -12.930 us | -10.57% |   FAIL   |
| I32 |    2^28    |        1         |        2         |       1       |   100000000   |  28.409 ms |       0.12% |  20.422 ms |       0.16% | -7987.039 us | -28.11% |   FAIL   |
| I32 |    2^14    |        10        |        2         |       1       |   100000000   | 127.440 us |       2.96% | 113.009 us |       4.54% |   -14.431 us | -11.32% |   FAIL   |
| I32 |    2^28    |        10        |        2         |       1       |   100000000   |  28.395 ms |       0.12% |  20.789 ms |       0.17% | -7606.146 us | -26.79% |   FAIL   |
| F64 |    2^14    |        1         |        2         |       1       |      10       | 119.900 us |       4.27% | 110.024 us |       5.21% |    -9.876 us |  -8.24% |   FAIL   |
| F64 |    2^28    |        1         |        2         |       1       |      10       |  26.547 ms |       0.14% |  18.790 ms |       0.19% | -7756.825 us | -29.22% |   FAIL   |
| F64 |    2^14    |        10        |        2         |       1       |      10       | 127.211 us |       3.62% | 112.724 us |       4.63% |   -14.487 us | -11.39% |   FAIL   |
| F64 |    2^28    |        10        |        2         |       1       |      10       |  30.087 ms |       0.31% |  23.886 ms |       0.19% | -6201.201 us | -20.61% |   FAIL   |
| F64 |    2^14    |        1         |        2         |       1       |      100      | 120.433 us |       2.42% | 109.777 us |       4.76% |   -10.656 us |  -8.85% |   FAIL   |
| F64 |    2^28    |        1         |        2         |       1       |      100      |  26.568 ms |       0.14% |  18.797 ms |       0.18% | -7770.905 us | -29.25% |   FAIL   |
| F64 |    2^14    |        10        |        2         |       1       |      100      | 126.702 us |       3.34% | 112.492 us |       4.27% |   -14.210 us | -11.22% |   FAIL   |
| F64 |    2^28    |        10        |        2         |       1       |      100      |  30.106 ms |       0.24% |  24.001 ms |       0.46% | -6105.382 us | -20.28% |   FAIL   |
| F64 |    2^14    |        1         |        2         |       1       |    1000000    | 119.557 us |       4.07% | 110.355 us |       5.38% |    -9.203 us |  -7.70% |   FAIL   |
| F64 |    2^28    |        1         |        2         |       1       |    1000000    |  26.906 ms |       0.15% |  19.126 ms |       0.18% | -7779.215 us | -28.91% |   FAIL   |
| F64 |    2^14    |        10        |        2         |       1       |    1000000    | 128.546 us |       3.89% | 113.429 us |       2.36% |   -15.117 us | -11.76% |   FAIL   |
| F64 |    2^28    |        10        |        2         |       1       |    1000000    |  30.898 ms |       0.18% |  24.737 ms |       0.45% | -6161.068 us | -19.94% |   FAIL   |
| F64 |    2^14    |        1         |        2         |       1       |   100000000   | 120.927 us |       4.36% | 109.903 us |       4.67% |   -11.024 us |  -9.12% |   FAIL   |
| F64 |    2^28    |        1         |        2         |       1       |   100000000   |  29.997 ms |       0.12% |  21.968 ms |       0.16% | -8028.453 us | -26.76% |   FAIL   |
| F64 |    2^14    |        10        |        2         |       1       |   100000000   | 127.789 us |       3.78% | 112.861 us |       3.81% |   -14.928 us | -11.68% |   FAIL   |
| F64 |    2^28    |        10        |        2         |       1       |   100000000   |  36.555 ms |       0.15% |  29.882 ms |       0.12% | -6672.173 us | -18.25% |   FAIL   |

# row_fixed_rolling_sum

## [0] NVIDIA RTX A6000

|  T  |  num_rows  |  preceding_size  |  following_size  |  min_periods  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |         Diff |   %Diff |  Status  |
|-----|------------|------------------|------------------|---------------|------------|-------------|------------|-------------|--------------|---------|----------|
| I32 |    2^14    |        1         |        2         |       1       |  31.751 us |       4.87% |  17.603 us |       6.85% |   -14.148 us | -44.56% |   FAIL   |
| I32 |    2^22    |        1         |        2         |       1       | 206.365 us |       1.51% |  95.830 us |       2.39% |  -110.534 us | -53.56% |   FAIL   |
| I32 |    2^28    |        1         |        2         |       1       |  10.883 ms |       0.20% |   4.767 ms |       0.25% | -6115.576 us | -56.20% |   FAIL   |
| I32 |    2^14    |        10        |        2         |       1       |  31.956 us |       5.91% |  17.559 us |       7.14% |   -14.397 us | -45.05% |   FAIL   |
| I32 |    2^22    |        10        |        2         |       1       | 206.654 us |       1.65% |  95.575 us |       2.29% |  -111.079 us | -53.75% |   FAIL   |
| I32 |    2^28    |        10        |        2         |       1       |  10.872 ms |       0.12% |   4.758 ms |       0.29% | -6114.139 us | -56.24% |   FAIL   |
| I32 |    2^14    |       100        |        2         |       1       |  40.611 us |       4.58% |  26.636 us |       9.61% |   -13.976 us | -34.41% |   FAIL   |
| I32 |    2^22    |       100        |        2         |       1       | 293.156 us |       1.54% | 235.445 us |       2.23% |   -57.710 us | -19.69% |   FAIL   |
| I32 |    2^28    |       100        |        2         |       1       |  17.109 ms |       0.42% |  14.613 ms |       0.48% | -2496.488 us | -14.59% |   FAIL   |
| I32 |    2^14    |        1         |        2         |      20       |  32.786 us |       6.05% |  18.115 us |      10.13% |   -14.671 us | -44.75% |   FAIL   |
| I32 |    2^22    |        1         |        2         |      20       | 206.530 us |       1.61% |  96.530 us |       2.79% |  -109.999 us | -53.26% |   FAIL   |
| I32 |    2^28    |        1         |        2         |      20       |  10.883 ms |       0.16% |   4.767 ms |       0.27% | -6115.283 us | -56.19% |   FAIL   |
| I32 |    2^14    |        10        |        2         |      20       |  31.927 us |       4.21% |  17.784 us |      15.89% |   -14.144 us | -44.30% |   FAIL   |
| I32 |    2^22    |        10        |        2         |      20       | 206.093 us |       1.58% |  95.657 us |       2.92% |  -110.436 us | -53.59% |   FAIL   |
| I32 |    2^28    |        10        |        2         |      20       |  10.872 ms |       0.12% |   4.757 ms |       0.06% | -6115.435 us | -56.25% |   FAIL   |
| I32 |    2^14    |       100        |        2         |      20       |  40.361 us |       3.87% |  26.359 us |       3.42% |   -14.002 us | -34.69% |   FAIL   |
| I32 |    2^22    |       100        |        2         |      20       | 292.868 us |       1.80% | 235.340 us |       2.69% |   -57.528 us | -19.64% |   FAIL   |
| I32 |    2^28    |       100        |        2         |      20       |  17.075 ms |       0.47% |  14.613 ms |       0.53% | -2462.145 us | -14.42% |   FAIL   |
| F64 |    2^14    |        1         |        2         |       1       |  35.335 us |      11.59% |  18.771 us |      13.65% |   -16.564 us | -46.88% |   FAIL   |
| F64 |    2^22    |        1         |        2         |       1       | 230.322 us |       1.42% | 121.412 us |       3.23% |  -108.910 us | -47.29% |   FAIL   |
| F64 |    2^28    |        1         |        2         |       1       |  12.460 ms |       0.09% |   6.311 ms |       0.11% | -6149.789 us | -49.35% |   FAIL   |
| F64 |    2^14    |        10        |        2         |       1       |  39.365 us |       8.03% |  24.712 us |      14.97% |   -14.653 us | -37.22% |   FAIL   |
| F64 |    2^22    |        10        |        2         |       1       | 283.241 us |       1.25% | 224.296 us |       1.86% |   -58.945 us | -20.81% |   FAIL   |
| F64 |    2^28    |        10        |        2         |       1       |  15.983 ms |       0.13% |  13.039 ms |       0.86% | -2943.223 us | -18.42% |   FAIL   |
| F64 |    2^14    |       100        |        2         |       1       |  46.514 us |       8.38% |  28.323 us |      10.77% |   -18.191 us | -39.11% |   FAIL   |
| F64 |    2^22    |       100        |        2         |       1       |   1.739 ms |       0.16% |   1.668 ms |       0.53% |   -70.890 us |  -4.08% |   FAIL   |
| F64 |    2^28    |       100        |        2         |       1       | 108.509 ms |       0.06% | 104.891 ms |       0.04% | -3618.662 us |  -3.33% |   FAIL   |
| F64 |    2^14    |        1         |        2         |      20       |  31.566 us |       6.77% |  17.406 us |       8.84% |   -14.160 us | -44.86% |   FAIL   |
| F64 |    2^22    |        1         |        2         |      20       | 230.307 us |       1.47% | 121.044 us |       1.90% |  -109.263 us | -47.44% |   FAIL   |
| F64 |    2^28    |        1         |        2         |      20       |  12.462 ms |       0.08% |   6.315 ms |       0.21% | -6147.508 us | -49.33% |   FAIL   |
| F64 |    2^14    |        10        |        2         |      20       |  39.277 us |       6.21% |  23.033 us |      16.63% |   -16.244 us | -41.36% |   FAIL   |
| F64 |    2^22    |        10        |        2         |      20       | 281.374 us |       1.09% | 219.307 us |       1.12% |   -62.067 us | -22.06% |   FAIL   |
| F64 |    2^28    |        10        |        2         |      20       |  15.944 ms |       0.47% |  12.882 ms |       0.72% | -3062.193 us | -19.21% |   FAIL   |
| F64 |    2^14    |       100        |        2         |      20       |  46.079 us |       8.03% |  28.287 us |      11.55% |   -17.792 us | -38.61% |   FAIL   |
| F64 |    2^22    |       100        |        2         |      20       |   1.740 ms |       0.19% |   1.664 ms |       0.36% |   -75.602 us |  -4.35% |   FAIL   |
| F64 |    2^28    |       100        |        2         |      20       | 108.475 ms |       0.04% | 104.920 ms |       0.04% | -3555.241 us |  -3.28% |   FAIL   |

# row_variable_rolling_sum

## [0] NVIDIA RTX A6000

|  T  |  num_rows  |  preceding_size  |  following_size  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |       Diff |   %Diff |  Status  |
|-----|------------|------------------|------------------|------------|-------------|------------|-------------|------------|---------|----------|
| I32 |    2^14    |        10        |        2         |  20.407 us |      17.95% |  17.981 us |       9.19% |  -2.427 us | -11.89% |   FAIL   |
| I32 |    2^22    |        10        |        2         | 145.353 us |       1.96% | 145.458 us |       1.82% |   0.105 us |   0.07% |   PASS   |
| I32 |    2^28    |        10        |        2         |   7.850 ms |       0.17% |   7.863 ms |       0.21% |  12.240 us |   0.16% |   PASS   |
| I32 |    2^14    |       100        |        2         |  26.130 us |       6.01% |  26.324 us |       3.86% |   0.194 us |   0.74% |   PASS   |
| I32 |    2^22    |       100        |        2         | 234.618 us |       2.86% | 245.878 us |       3.30% |  11.260 us |   4.80% |   FAIL   |
| I32 |    2^28    |       100        |        2         |  14.624 ms |       0.79% |  15.221 ms |       0.68% | 596.983 us |   4.08% |   FAIL   |
| F64 |    2^14    |        10        |        2         |  26.256 us |       2.86% |  26.508 us |       6.46% |   0.252 us |   0.96% |   PASS   |
| F64 |    2^22    |        10        |        2         | 223.001 us |       1.96% | 222.131 us |       1.85% |  -0.870 us |  -0.39% |   PASS   |
| F64 |    2^28    |        10        |        2         |  12.896 ms |       0.58% |  12.893 ms |       0.78% |  -3.356 us |  -0.03% |   PASS   |
| F64 |    2^14    |       100        |        2         |  33.438 us |       6.71% |  34.323 us |       1.75% |   0.885 us |   2.65% |   FAIL   |
| F64 |    2^22    |       100        |        2         |   1.667 ms |       0.49% |   1.672 ms |       0.26% |   5.138 us |   0.31% |   FAIL   |
| F64 |    2^28    |       100        |        2         | 104.673 ms |       0.05% | 104.982 ms |       0.04% | 309.519 us |   0.30% |   FAIL   |

# Summary

- Total Matches: 80
  - Pass    (diff <= min_noise): 6
  - Unknown (infinite noise):    0
  - Failure (diff > min_noise):  74

</details>

<details> <summary> Before </summary>

# Benchmark Results

## row_grouped_rolling_sum

### [0] NVIDIA RTX A6000

|  T  |     num_rows     | preceding_size | following_size | min_periods | cardinality | Samples |  CPU Time  | Noise |  GPU Time  | Noise | Mrows/s | peak_memory_usage |
|-----|------------------|----------------|----------------|-------------|-------------|---------|------------|-------|------------|-------|---------|-------------------|
| I32 |     2^14 = 16384 |              1 |              2 |           1 |          10 |    840x | 127.585 us | 5.09% | 123.556 us | 3.90% |     132 |       450.031 KiB |
| I32 | 2^28 = 268435456 |              1 |              2 |           1 |          10 |    601x |  24.948 ms | 0.13% |  24.945 ms | 0.13% |   10761 |         7.031 GiB |
| I32 |     2^14 = 16384 |             10 |              2 |           1 |          10 |   1020x | 126.638 us | 5.18% | 122.737 us | 4.09% |     133 |       450.031 KiB |
| I32 | 2^28 = 268435456 |             10 |              2 |           1 |          10 |    601x |  24.947 ms | 0.18% |  24.944 ms | 0.18% |   10761 |         7.031 GiB |
| I32 |     2^14 = 16384 |              1 |              2 |           1 |         100 |    630x | 123.913 us | 5.30% | 119.947 us | 4.07% |     136 |       450.031 KiB |
| I32 | 2^28 = 268435456 |              1 |              2 |           1 |         100 |    601x |  24.958 ms | 0.15% |  24.954 ms | 0.15% |   10757 |         7.031 GiB |
| I32 |     2^14 = 16384 |             10 |              2 |           1 |         100 |    984x | 127.839 us | 4.98% | 123.919 us | 3.85% |     132 |       450.031 KiB |
| I32 | 2^28 = 268435456 |             10 |              2 |           1 |         100 |    601x |  24.958 ms | 0.14% |  24.955 ms | 0.14% |   10756 |         7.031 GiB |
| I32 |     2^14 = 16384 |              1 |              2 |           1 |     1000000 |    802x | 125.851 us | 5.65% | 121.896 us | 4.59% |     134 |       450.031 KiB |
| I32 | 2^28 = 268435456 |              1 |              2 |           1 |     1000000 |    593x |  25.294 ms | 0.15% |  25.291 ms | 0.15% |   10613 |         7.031 GiB |
| I32 |     2^14 = 16384 |             10 |              2 |           1 |     1000000 |    992x | 130.416 us | 4.81% | 126.450 us | 3.65% |     129 |       450.031 KiB |
| I32 | 2^28 = 268435456 |             10 |              2 |           1 |     1000000 |    593x |  25.298 ms | 0.14% |  25.295 ms | 0.13% |   10612 |         7.031 GiB |
| I32 |     2^14 = 16384 |              1 |              2 |           1 |   100000000 |    820x | 126.300 us | 5.56% | 122.268 us | 4.42% |     134 |       450.031 KiB |
| I32 | 2^28 = 268435456 |              1 |              2 |           1 |   100000000 |    528x |  28.413 ms | 0.12% |  28.409 ms | 0.12% |    9448 |         7.031 GiB |
| I32 |     2^14 = 16384 |             10 |              2 |           1 |   100000000 |    752x | 131.420 us | 4.30% | 127.440 us | 2.96% |     128 |       450.031 KiB |
| I32 | 2^28 = 268435456 |             10 |              2 |           1 |   100000000 |    528x |  28.398 ms | 0.12% |  28.395 ms | 0.12% |    9453 |         7.031 GiB |
…