@wence- 2024-12-18
Sits on top of #17613, so don't merge until that is in and this is rebased/has trunk back-merged.