@wence- 2024-12-18
Note that I haven't touched the jit rolling window kernel yet (some of the same changes will apply there eventually).

@mythrocks This changes the API contract for to rolling_window, so spark may need to adapt.