@mythrocks 2024-12-18
> @mythrocks This changes the API contract for to rolling_window, so spark may need to adapt.

I'm trying to work out what of this has impact on `spark-rapids`.  Spark calls into `cudf::grouped_rolling*()` with fixed preceding/following offsets, which might exceed group bounds at the margins.