@wence- 2024-12-19
> > @mythrocks This changes the API contract for to rolling_window, so spark may need to adapt.
> 
> I'm trying to work out what of this has impact on `spark-rapids`. Spark calls into `cudf::grouped_rolling*()` with fixed preceding/following offsets, which might exceed group bounds at the margins.

I think this is a call into https://github.com/rapidsai/cudf/blob/030ae7af940a734963aa5f99e1203e2b2cc0c55b/cpp/src/rolling/grouped_rolling.cu#L47-L55

If so, those fixed offsets are still clamped to the group bounds, via a transform iterator:

https://github.com/rapidsai/cudf/blob/030ae7af940a734963aa5f99e1203e2b2cc0c55b/cpp/src/rolling/grouped_rolling.cu#L127-L133