@mythrocks 2024-12-19
> If so, those fixed offsets are still clamped to the group bounds, via a transform iterator:

That's what I thought.  It seems to me like this really ought to work for `spark-rapids` without code change, if the transform iterator's doing the clamping.

I'll try have a go at integrating this with `spark-rapids`, and running some tests.