@mythrocks 2025-01-06
I'd like to double-check my work, but it appears that the window-function tests in `spark-rapids` are running correctly with this refactored version.

I'll have another go at the tests to confirm.